made fixes

+++ src/main/java/io/codelex/advancedTest/cards/Card.java
@@ -1,24 +1,26 @@
 package io.codelex.advancedTest.cards;
 
+import java.math.BigDecimal;
+
 public abstract class Card {
 
   private int number;
   private String fullName;
   private String ccv;
-  private double balance;
+  private BigDecimal balance;
 
   public Card(int number, String fullName, String ccv, double balance) {
     this.number = number;
     this.fullName = fullName;
     this.ccv = ccv;
-    this.balance = balance;
+    this.balance = BigDecimal.valueOf(balance);
   }
 
-  public double getBalance() {
+  public BigDecimal getBalance() {
     return balance;
   }
 
-  public void setBalance(double balance) {
+  public void setBalance(BigDecimal balance) {
     this.balance = balance;
   }
